tests: Move MessageHasKeywordsTest to test_message_fetch.py.
